CARDS-1608: More helpful page displayed to non-patient users trying to access /Proms.html
